# hash algorithm improvements