--- a/std/hash.zig
+++ b/std/hash.zig
@@ -16,12 +16,15 @@ pub const SipHash128 = siphash.SipHash128;
 
 pub const murmur = @import("hash/murmur.zig");
 pub const Murmur2_32 = murmur.Murmur2_32;
+
 pub const Murmur2_64 = murmur.Murmur2_64;
 pub const Murmur3_32 = murmur.Murmur3_32;
 
 pub const cityhash = @import("hash/cityhash.zig");
 pub const CityHash32 = cityhash.CityHash32;
 pub const CityHash64 = cityhash.CityHash64;
+
+pub const wyhash = @import("hash/wyhash.zig").hash;
 
 test "hash" {
     _ = @import("hash/adler.zig");
@@ -30,4 +33,5 @@ test "hash" {
     _ = @import("hash/siphash.zig");
     _ = @import("hash/murmur.zig");
     _ = @import("hash/cityhash.zig");
+    _ = @import("hash/wyhash.zig");
 }