Closes #2244.