@tiehuis 2019-07-01
I'll make a PR against your branch shortly with some code for testing the speed of all the hash functions. Would be nice to have some measurements here before merging to justify the change. Actually, this will need to have an iterative api as well so I'll see what I can do there, too.