@Sahnvour 2019-07-02
I won't have time to work again on this until mid-july, but I still want to investigate performance and "de-hardcoding" the core function of wyhash as @tiehuis proposed.
It's also blocking at the moment on #2604 for `autoHash`'s union implementation.