@Sahnvour 2019-07-16
@tiehuis The runtime version is indeed slower for small keys. My benchmarks on hashmaps with `u64` keys got around 20% slower, so probably a lot more if only considering the hashing itself.

I also removed the iterative accumulation of small inputs as it proved ~15% slower on my machine.

### with internal state
```
zig run ./std/hash/throughput_test.zig --release-fast --override-std-dir ./std/ -- --seed 0 --count 1024

      wyhash: 9550 MiB/s [5532736999793325286]
siphash(1,3): 3864 MiB/s [4293770063772332796]
siphash(2,4): 2070 MiB/s [4826931557822930874]
       fnv1a: 0940 MiB/s [11215122837159944997]
       crc32: 2409 MiB/s [0000001171707108]
```

### without internal state
```
zig run ./std/hash/throughput_test.zig --release-fast --override-std-dir ./std/ -- --seed 0 --count 1024

      wyhash: 11167 MiB/s [5532736999793325286]
siphash(1,3): 3864 MiB/s [4293770063772332796]
siphash(2,4): 2070 MiB/s [4826931557822930874]
       fnv1a: 0941 MiB/s [11215122837159944997]
       crc32: 2393 MiB/s [0000001171707108]

zig run ./std/hash/throughput_test.zig --release-safe --override-std-dir ./std/ -- --seed 0 --count 1024

      wyhash: 11075 MiB/s [5532736999793325286]
siphash(1,3): 3862 MiB/s [4293770063772332796]
siphash(2,4): 1355 MiB/s [4826931557822930874]
       fnv1a: 0941 MiB/s [11215122837159944997]
       crc32: 2387 MiB/s [0000001171707108]

zig run ./std/hash/throughput_test.zig --release-small --override-std-dir ./std/ -- --seed 0 --count 1024

      wyhash: 1136 MiB/s [5532736999793325286]
siphash(1,3): 3777 MiB/s [4293770063772332796]
siphash(2,4): 2029 MiB/s [4826931557822930874]
       fnv1a: 0941 MiB/s [11215122837159944997]
       crc32: 2314 MiB/s [0000001171707108]

zig run ./std/hash/throughput_test.zig --override-std-dir ./std/ -- --seed 0 --count 1024

      wyhash: 0154 MiB/s [5532736999793325286]
siphash(1,3): 0058 MiB/s [4293770063772332796]
siphash(2,4): 0030 MiB/s [4826931557822930874]
       fnv1a: 0236 MiB/s [11215122837159944997]
       crc32: 0321 MiB/s [0000001171707108]
```